Better tests.
